Change:VIRTIO0 i32 to usize

+++ kernel-rs/src/memlayout.rs
@@ -22,7 +22,7 @@ pub const UART0: usize = 0x10000000;
 pub const UART0_IRQ: usize = 10;
 
 /// virtio mmio interface
-pub const VIRTIO0: i32 = 0x10001000;
+pub const VIRTIO0: usize = 0x10001000;
 pub const VIRTIO0_IRQ: i32 = 1;
 
 /// local interrupt controller, which contains the timer.
